+++ versions/0_4_rc/flake.nix
@@ -2,7 +2,7 @@
   inputs =
     {
       holochain = {
-        url = "github:holochain/holochain/holochain-0.5.0-dev.0";
+        url = "github:holochain/holochain/holochain-0.4.0-rc.0";
         flake = false;
       };
 
@@ -13,13 +13,13 @@
 
       # holochain_cli_launch
       launcher = {
-        url = "github:holochain/hc-launch/holochain-weekly";
+        url = "github:holochain/hc-launch/holochain-0.4";
         flake = false;
       };
 
       # holochain_scaffolding_cli
       scaffolding = {
-        url = "github:holochain/scaffolding/holochain-weekly";
+        url = "github:holochain/scaffolding/holochain-0.4";
         flake = false;
       };
     };